docs: generate documentation with swagger ui

+++ src/use-cases/project/delete-project.ts
@@ -1,4 +1,5 @@
 import { ProjectRepository } from '@/infra/db/protocols';
+import { InvalidParamError } from '@/util/errors';
 
 export class DeleteProjectUseCase {
   private projectRepository: ProjectRepository;
@@ -8,6 +9,10 @@ export class DeleteProjectUseCase {
   }
 
   async execute(id: string): Promise<void> {
+    const projectExists = await this.projectRepository.getProject(id);
+    if (!projectExists) {
+      throw new InvalidParamError('projectId');
+    }
     await this.projectRepository.deleteProject(id);
   }
 }
